Add save functionality to the game. Incomplete
Create a save method on the game class. The game object will be serialized
(YAML/PSYCH) and written in saved_games.txt. The game is not able to read
it or load it yet

Modified take guess method on game class. That way the user gets to pick
a letter or "save the game" or "quit"