Save/Load capabilities added to the game
The user is now asked to start a new game or save a new game. The if loop
inside the game_on loop allows the user to be prompted for a 'new' or 'load'
game after the round finishes.

Create load_game method on game class that reads the YAML object stored.

Create new_game message and load_game message when the game starts.
(Display class)

Improve spacing in Display methods.